backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
python-multipart
passlib[bcrypt]
bcrypt==4.0.1